requirements.txt
cryptography>=3.0
pygame>=2.0.0.dev10
pillow>=7.2.0
twisted>=20.3.0